Создаст класс `board-controller`
